+++ projects/hiptensor/cmake/Functions/hiptensorSupportedArchitectures.cmake
@@ -6,6 +6,7 @@ set(SUPPORTED_ARCHITECTURES
     gfx1100
     gfx1101
     gfx1102
+    gfx1103
     gfx1150
     gfx1151
     gfx1152