+++ projects/hiptensor/library/src/hip_device.cpp
@@ -74,6 +74,10 @@ namespace hiptensor
         {
             mGcnArch = hipGcnArch_t::GFX1102;
         }
+        else if(deviceName.find("gfx1103") != std::string::npos)
+        {
+            mGcnArch = hipGcnArch_t::GFX1103;
+        }
         else if(deviceName.find("gfx1150") != std::string::npos)
         {
             mGcnArch = hipGcnArch_t::GFX1150;
@@ -158,6 +162,7 @@ namespace hiptensor
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1100
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1101
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1102
+                    || mGcnArch == HipDevice::hipGcnArch_t::GFX1103
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1150
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1151
                     || mGcnArch == HipDevice::hipGcnArch_t::GFX1152